Cleaned up the code. Split up Import and Export Python files. Export works now.
