+++ test/plugins/genbank.spec.js
@@ -1,7 +1,7 @@
-import { expect } from 'chai';
+import {expect} from 'chai';
 import path from 'path';
-import { importProject, importConstruct } from '../../plugins/convert/import';
-import { exportProject } from '../../plugins/convert/export';
+import {importProject, importConstruct} from '../../plugins/convert/import';
+import {exportProject} from '../../plugins/convert/export';
 import BlockDefinition from '../../src/schemas/Block';
 import ProjectDefinition from '../../src/schemas/Project';
 
@@ -13,43 +13,43 @@ const getBlock = (allBlocks, blockId) => {
       return allBlocks[i];
     }
   }
-}
+};
 
 describe('Plugins', () => {
   describe.only('Genbank Plugin', () => {
     it('should import Genbank file with contiguous entries as a project', function importGB(done) {
-        fs.readFile(path.resolve(__dirname, '../res/sampleGenbankContiguous.gb'), 'utf8', (err, sampleStr) => {
-            importProject('genbank', sampleStr)
-                .then(output => {
-                    expect(output.project).not.to.equal(undefined);
-                    expect(ProjectDefinition.validate(output.project)).to.equal(true);
-                    expect(output.project.metadata.name).to.equal('EU912544');
-                    expect(output.project.metadata.description).to.equal('Cloning vector pDM313, complete sequence.')
-                    expect(output.project.components.length).to.equal(1);
-                    const parentBlock = getBlock(output.blocks, output.project.components[0]);
-                    expect(parentBlock.components.length).to.equal(4);
-                    expect(parentBlock.metadata.name).to.equal('EU912544');
-                    expect(parentBlock.metadata.description).to.equal('Cloning vector pDM313, complete sequence.');
-                    expect(parentBlock.metadata.genbank.mol_type).to.equal('other DNA');
-                    expect(parentBlock.metadata.start).to.equal(0);
-                    expect(parentBlock.metadata.end).to.equal(119);
-                    expect(parentBlock.metadata.genbank.annotations.gi).to.equal('198078160');
-                    expect(parentBlock.metadata.genbank.note).to.equal('GFP-tag for C-terminal fusion');
-                    expect(parentBlock.metadata.genbank.annotations.data_file_division).to.equal('SYN');
-                    expect(parentBlock.metadata.genbank.annotations.date).to.equal('06-FEB-2009');
-                    expect(getBlock(output.blocks, parentBlock.components[0]).metadata.type).to.equal('promoter');
-                    expect(getBlock(output.blocks, parentBlock.components[0]).rules.sbol).to.equal('promoter');
-                    expect(getBlock(output.blocks, parentBlock.components[1]).metadata.type).to.equal('CDS');
-                    expect(getBlock(output.blocks, parentBlock.components[1]).rules.sbol).to.equal('cds');
-                    expect(getBlock(output.blocks, parentBlock.components[2]).metadata.type).to.equal('terminator');
-                    expect(getBlock(output.blocks, parentBlock.components[3]).metadata.type).to.equal('rep_origin');
-                    for (var key in output.blocks) {
-                      expect(BlockDefinition.validate(output.blocks[key])).to.equal(true);
-                    }
-                    done();
-                })
-                .catch(done);
-        });
+      fs.readFile(path.resolve(__dirname, '../res/sampleGenbankContiguous.gb'), 'utf8', (err, sampleStr) => {
+        importProject('genbank', sampleStr)
+          .then(output => {
+            expect(output.project).not.to.equal(undefined);
+            expect(ProjectDefinition.validate(output.project)).to.equal(true);
+            expect(output.project.metadata.name).to.equal('EU912544');
+            expect(output.project.metadata.description).to.equal('Cloning vector pDM313, complete sequence.');
+            expect(output.project.components.length).to.equal(1);
+            const parentBlock = getBlock(output.blocks, output.project.components[0]);
+            expect(parentBlock.components.length).to.equal(4);
+            expect(parentBlock.metadata.name).to.equal('EU912544');
+            expect(parentBlock.metadata.description).to.equal('Cloning vector pDM313, complete sequence.');
+            expect(parentBlock.metadata.genbank.feature_annotations.mol_type).to.equal('other DNA');
+            expect(parentBlock.metadata.start).to.equal(0);
+            expect(parentBlock.metadata.end).to.equal(119);
+            expect(parentBlock.metadata.genbank.annotations.gi).to.equal('198078160');
+            expect(parentBlock.metadata.genbank.feature_annotations.note).to.equal('GFP-tag for C-terminal fusion');
+            expect(parentBlock.metadata.genbank.annotations.data_file_division).to.equal('SYN');
+            expect(parentBlock.metadata.genbank.annotations.date).to.equal('06-FEB-2009');
+            expect(getBlock(output.blocks, parentBlock.components[0]).metadata.type).to.equal('promoter');
+            expect(getBlock(output.blocks, parentBlock.components[0]).rules.sbol).to.equal('promoter');
+            expect(getBlock(output.blocks, parentBlock.components[1]).metadata.type).to.equal('CDS');
+            expect(getBlock(output.blocks, parentBlock.components[1]).rules.sbol).to.equal('cds');
+            expect(getBlock(output.blocks, parentBlock.components[2]).metadata.type).to.equal('terminator');
+            expect(getBlock(output.blocks, parentBlock.components[3]).metadata.type).to.equal('rep_origin');
+            for (let key in output.blocks) {
+              expect(BlockDefinition.validate(output.blocks[key])).to.equal(true);
+            }
+            done();
+          })
+          .catch(done);
+      });
     });
 
     it('should import Genbank file with holes as a project', function importGB(done) {
@@ -67,9 +67,9 @@ describe('Plugins', () => {
             expect(getBlock(output.blocks, parentBlock.components[4]).metadata.type).to.equal('filler');
             expect(getBlock(output.blocks, parentBlock.components[5]).metadata.type).to.equal('terminator');
             expect(getBlock(output.blocks, parentBlock.components[6]).metadata.type).to.equal('rep_origin');
-            for (var key in output.blocks) {
+            for (let key in output.blocks) {
               expect(BlockDefinition.validate(output.blocks[key])).to.equal(true);
-            };
+            }
             done();
           })
           .catch(done);
@@ -84,22 +84,22 @@ describe('Plugins', () => {
             expect(output.project.components.length === 1).to.equal(true);
             const parentBlock = getBlock(output.blocks, output.project.components[0]);
             expect(parentBlock.components.length).to.equal(2);
-            var firstBlock = getBlock(output.blocks, parentBlock.components[0]);
+            let firstBlock = getBlock(output.blocks, parentBlock.components[0]);
             expect(firstBlock.metadata.type).to.equal('block');
             expect(firstBlock.components.length).to.be.equal(3);
             expect(getBlock(output.blocks, firstBlock.components[0]).metadata.type).to.equal('promoter');
             expect(getBlock(output.blocks, firstBlock.components[1]).metadata.type).to.equal('CDS');
             expect(getBlock(output.blocks, firstBlock.components[2]).metadata.type).to.equal('filler');
-            var secondBlock = getBlock(output.blocks, parentBlock.components[1]);
+            let secondBlock = getBlock(output.blocks, parentBlock.components[1]);
             expect(secondBlock.metadata.type).to.equal('block');
             expect(secondBlock.components.length).to.be.equal(4);
             expect(getBlock(output.blocks, secondBlock.components[0]).metadata.type).to.equal('CDS');
             expect(getBlock(output.blocks, secondBlock.components[1]).metadata.type).to.equal('filler');
             expect(getBlock(output.blocks, secondBlock.components[2]).metadata.type).to.equal('terminator');
             expect(getBlock(output.blocks, secondBlock.components[3]).metadata.type).to.equal('rep_origin');
-            for (var key in output.blocks) {
+            for (let key in output.blocks) {
               expect(BlockDefinition.validate(output.blocks[key])).to.equal(true);
-            };
+            }
             done();
           })
           .catch(done);
@@ -132,22 +132,22 @@ describe('Plugins', () => {
             expect(output.roots.length).to.equal(1);
             const parentBlock = getBlock(output.blocks, output.roots[0]);
             expect(parentBlock.components.length).to.equal(2);
-            var firstBlock = getBlock(output.blocks, parentBlock.components[0]);
+            let firstBlock = getBlock(output.blocks, parentBlock.components[0]);
             expect(firstBlock.metadata.type).to.equal('block');
             expect(firstBlock.components.length).to.be.equal(3);
             expect(getBlock(output.blocks, firstBlock.components[0]).metadata.type).to.equal('promoter');
             expect(getBlock(output.blocks, firstBlock.components[1]).metadata.type).to.equal('CDS');
             expect(getBlock(output.blocks, firstBlock.components[2]).metadata.type).to.equal('filler');
-            var secondBlock = getBlock(output.blocks, parentBlock.components[1]);
+            let secondBlock = getBlock(output.blocks, parentBlock.components[1]);
             expect(secondBlock.metadata.type).to.equal('block');
             expect(secondBlock.components.length).to.be.equal(4);
             expect(getBlock(output.blocks, secondBlock.components[0]).metadata.type).to.equal('CDS');
             expect(getBlock(output.blocks, secondBlock.components[1]).metadata.type).to.equal('filler');
             expect(getBlock(output.blocks, secondBlock.components[2]).metadata.type).to.equal('terminator');
             expect(getBlock(output.blocks, secondBlock.components[3]).metadata.type).to.equal('rep_origin');
-            for (var key in output.blocks) {
+            for (let key in output.blocks) {
               expect(BlockDefinition.validate(output.blocks[key])).to.equal(true);
-            };
+            }
             done();
           })
           .catch(done);
@@ -170,9 +170,40 @@ describe('Plugins', () => {
         const sampleProject = JSON.parse(sampleProjectJson);
         exportProject('genbank', sampleProject)
           .then(result => {
-            console.log('came back!');
             expect(result).to.contain('LOCUS');
             done();
+          })
+          .catch(done);
+      });
+    });
+
+    it('should roundtrip a Genbank through our app', function exportGB(done) {
+      fs.readFile(path.resolve(__dirname, '../res/sampleGenbankContiguous.gb'), 'utf8', (err, sampleStr) => {
+        importProject('genbank', sampleStr)
+          .then(output => {
+            expect(output.project).not.to.equal(undefined);
+            expect(ProjectDefinition.validate(output.project)).to.equal(true);
+            expect(output.project.metadata.name).to.equal('EU912544');
+            expect(output.project.metadata.description).to.equal('Cloning vector pDM313, complete sequence.')
+            exportProject('genbank', output)
+              .then(result => {
+                expect(result).to.contain('LOCUS       EU912544                 120 bp    DNA');
+                expect(result).to.contain('SYN 06-FEB-2009');
+                expect(result).to.contain('DEFINITION  Cloning vector pDM313, complete sequence.');
+                expect(result).to.contain('ACCESSION   EU912544');
+                expect(result).to.contain('VERSION     EU912544.1  GI:198078160');
+                expect(result).to.contain('SOURCE      Cloning vector pDM313');
+                expect(result).to.contain('ORGANISM  Cloning vector pDM313');
+                expect(result).to.contain('other sequences; artificial sequences; vectors.');
+                expect(result).to.contain('REFERENCE   1');
+                expect(result).to.contain('AUTHORS   Veltman,D.M., Akar,G., Bosgraaf,L. and Van Haastert,P.J.');
+                expect(result).to.contain('TITLE     A new set of small, extrachromosomal expression vectors for');
+                expect(result).to.contain('Dictyostelium discoideum');
+                expect(result).to.contain('JOURNAL   Plasmid 61 (2), 110-118 (2009)');
+                expect(result).to.contain('PUBMED   19063918');
+                expect(result).to.contain('');
+                done();
+              });
           })
           .catch(done);
       });
